Warn when xnack disabled, move xnack check impl to .cpp
