+++ core/src/HIP/Kokkos_HIP.cpp
@@ -75,12 +75,13 @@ void HIP::impl_initialize(InitializationSettings const& settings) {
   }
 #endif
 #ifdef KOKKOS_ARCH_AMD_GFX942_APU
-  if (!Impl::HIPInternal::singleton().is_xnack_enabled()) {
-    Kokkos::abort(
-        "Could not determine that xnack is enabled. Kokkos requires xnack to "
-        "be enabled for ARCH_AMD_GFX942_APU (MI300A). Set HSA_XNACK=1 in your "
-        "environment and ensure \"CONFIG_HMM_MIRROR=y\" is in the /boot/config "
-        "file and that file is readable\n");
+  if (!Kokkos::Impl::xnack_enabled()) {
+    std::cerr << R"warning(
+Kokkos::HIP::initialize WARNING: Could not determine that xnack is enabled. 
+                                 Kokkos requires xnack to be enabled for ARCH_AMD_GFX942_APU (MI300A).
+                                 Set HSA_XNACK=1 in your environment and ensure
+                                 \"CONFIG_HMM_MIRROR=y\" is in the /boot/config file and that file is readable)warning"
+              << std::endl;
   }
 
   if ((Kokkos::show_warnings()) &&
