+++ core/src/HIP/Kokkos_HIP_IsXnack.hpp
@@ -19,75 +19,31 @@
 #ifndef KOKKOS_HIP_ISXNACK_HPP
 #define KOKKOS_HIP_ISXNACK_HPP
 
-#include <string>
-#include <optional>
-#include <fstream>
-
-#ifdef __linux__
-#include <stdio.h>
-#include <sys/utsname.h>
-#endif
-
 namespace Kokkos::Impl {
 
-inline bool is_hsa_xnack_1_impl() {
-  const char* envVar = std::getenv("HSA_XNACK");
-  return envVar != nullptr && std::string(envVar) == "1";
-}
+/* Returns true iff we think the AMD GPU can access allocations created with the
+system allocator.
 
-// returns true iff environment variable HSA_XNACK=1
-inline bool is_hsa_xnack_1() {
-  static bool cache = is_hsa_xnack_1_impl();
-  return cache;
-}
+Based on AMD's ROCm 6.3.1 documentation:
+https://github.com/ROCm/HIP/blob/2c240cacff16c2bb18ce9e5b4c1b937ab17a0199/docs/how-to/hip_runtime_api/memory_management/unified_memory.rst?plain=1#L141-L146
 
-#ifdef __linux__
-// try to get `uname -r`. Returns an empty optional for any problem
-inline std::optional<std::string> uname_r() {
-  struct utsname buffer;
-  if (uname(&buffer) != 0) {
-    return std::optional<std::string>{};
-  }
-  return std::optional<std::string>{buffer.release};
-}
-#endif
+    To ensure the proper functioning of system allocated unified memory on
+    supported GPUs, it is essential to configure the environment variable
+    ``XNACK=1`` (sic) and use a kernel that supports HMM. Without this
+    configuration, the behavior will be similar to that of systems without HMM
+    support.
 
-inline bool is_boot_config_hmm_mirror_y_impl() {
-#ifdef __linux__
-  // figure out the boot config file name
-  std::optional<std::string> unameR = uname_r();
-  if (!unameR) {
-    // couldn't figure out linux release name
-    return false;
-  }
-  std::string bootConfigPath = std::string("/boot/config-") + unameR.value();
+This clearly states two things are required:
+* HSA_XNACK=1 is set in the environment
+* The kernel must support HMM
 
-  std::ifstream file(bootConfigPath);
-  if (!file.is_open()) {
-    // couldn't open file for whatever reason
-    return false;
-  }
-
-  std::string line;
-  while (std::getline(file, line)) {
-    if (line.find("CONFIG_HMM_MIRROR=y") != std::string::npos) {
-      return true;  // Found the string
-    }
-  }
-  return false;
-
-#else
-  return false;
-#endif
-}
-
-// return true iff `CONFIG_HMM_MIRROR=y` is definitely in /boot/config-$(uname
-// -r) returns false for non-linux platforms or any other problem
-inline bool is_boot_config_hmm_mirror_y() {
-  static bool cache = is_boot_config_hmm_mirror_y_impl();
-  return cache;
-}
-
+Across a couple Nvidia and AMD systems, we have observed that
+CONFIG_HMM_MIRROR=y was set in /boot/config-$(uname -r). This test may need to
+be modified if a better way is determined to check for HMM support in Linux.
+Checking for CONFIG_HMM was considered, but it was not present on El Capitan,
+so we infer its presence is not necessary.
+*/
+bool xnack_enabled();
 }  // namespace Kokkos::Impl
 
 #endif  // KOKKOS_HIP_ISXNACK_HPP
